requirements.txt
click==8.3.0
cmake==4.1.0
colorama==0.4.6
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.3
six==1.17.0
tzdata==2025.2
streamlit==1.40.0